+++ packages/worker/CHANGELOG.md
@@ -10,4 +10,10 @@ First stable release.
 
 ## 1.0.1
 
-Improved documentation, added more examples.
+- Improved documentation.
+- Added more examples.
+
+## 1.1.0
+
+- Added `transferList` as a second argument of a task function.
+- Added examples for shared data and transfer list.